* Added AudioManager::SetMusicStreaming(bool) - defaults to on (same as before) but off will now cause music to be fully loaded and cached, sometimes better if you're switching back and forth a lot * (OSX) BUGFIX: Mouse wheel events are now floating point accuracy instead of being cast into an int * Added MESSAGE_TYPE_PLAY_MUSIC_FORCE_STREAMING, like MESSAGE_TYPE_PLAY_MUSIC but.. yep, forces streaming.  Useful for something I was doing

+++ shared/OSX/app/MyOpenGLView.mm
@@ -283,7 +283,7 @@ static CVReturn MyDisplayLinkCallback(CVDisplayLinkRef displayLink, const CVTime
 
 - (void)scrollWheel:(NSEvent *)theEvent
 {
-    int deltaY = [theEvent deltaY];
+    float deltaY = [theEvent deltaY];
     GetMessageManager()->SendGUIEx2(MESSAGE_TYPE_GUI_MOUSEWHEEL, (float)deltaY, 0, 0, 0);
 }
